chore: hook 파일에서 textarea의 text 분리
